+++ packages/server/src/server/handlers.ts
@@ -5,6 +5,6 @@ export * as network from './handlers/network';
 export * as telemetry from './handlers/telemetry';
 export * as tools from './handlers/tools';
 export * as workflows from './handlers/workflows';
-export * as vNextWorkflow from './handlers/vNextWorkflow';
+export * as vNextWorkflows from './handlers/vNextWorkflows';
 export * as vector from './handlers/vector';
 export * as voice from './handlers/voice';
